move json formatting tests to formatting test file
